migration

+++ database/migrations/2021_01_04_021444_create_jembatan_foto.php
@@ -13,14 +13,16 @@ class CreateJembatanFoto extends Migration
      */
     public function up()
     {
-        Schema::create('master_jembatan_foto', function (Blueprint $table) {
-            $table->id();
-            $table->integer('id_jembatan')->unsigned();
-            $table->string('nama', 100);
-            $table->text('foto');
-            $table->foreign('id_jembatan')->references('id')->on('master_jembatan')->onUpdate('cascade')
-                ->onDelete('cascade');;
-        });
+        if (!Schema::hasTable('master_jembatan_foto')) {
+            Schema::create('master_jembatan_foto', function (Blueprint $table) {
+                $table->id();
+                $table->integer('id_jembatan')->unsigned();
+                $table->string('nama', 100);
+                $table->text('foto');
+                $table->foreign('id_jembatan')->references('id')->on('master_jembatan')->onUpdate('cascade')
+                    ->onDelete('cascade');;
+            });
+        }
     }
 
     /**
